+++ src/demo_substrate_events/types/assethub/substrate_events/assets_transferred/__init__.py
@@ -1,4 +1,3 @@
 from .v601 import V601
-from .v700 import V700
 
-type AssetsTransferredPayload = V601 | V700
+type AssetsTransferredPayload = V601