+++ src/dipdup/fields.py
@@ -130,9 +130,7 @@ class DecimalField(Field[Decimal], Decimal):
         self.quant = Decimal('1' if decimal_places == 0 else f'1.{("0" * decimal_places)}')
 
     def to_python_value(self, value: Any) -> Decimal | None:
-        if value is None:
-            value = None
-        else:
+        if value is not None:
             value = Decimal(value).quantize(self.quant).normalize()
         self.validate(value)
         return value  # type: ignore[no-any-return]